Update OAC_T1.asm
The objective of the work is to practice RISC-V assembly programming, focusing on implementing part of the IDEA (International Data Encryption Algorithm) data encryption algorithm. IDEA is a symmetric encryption with a 128-bit key that operates in 64-bit blocks, divided into 16-bit sub-blocks. It
is based on three basic operations: exclusive-or (XOR) over 16 bits, modular sum (2^16), and modular multiplication
(2^(16+1)).